ready to deploy

+++ enrollment/api/handlers/imports.py
@@ -11,7 +11,7 @@ from .base import BaseView
 from .validation import ImportsSchema
 
 from sqlalchemy import and_
-from sqlalchemy import select, update
+from sqlalchemy import select, update, exists, and_, insert
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from enrollment.db.schema import Item, Relation
@@ -28,34 +28,38 @@ def create_item(data: dict, date: datetime) -> Item:
         size=data['size']
     )
 
+async def update_parents_recursive(session: AsyncSession, id: Union[str, None], date: datetime) -> List[str]:
+    updated = []
+    
+    if id is not None:
+        updated.append(id)
+
+        await session.execute(update(Item).where(Item.id == id).values(date=date))
+        next_item = await session.execute(select(Item.parentId).where(Item.id == id))
+        next_item = next_item.scalars().first()
+
+        updated.extend(await update_parents_recursive(session, next_item, date))
+
+    return updated
+
+
 class ImportsView(BaseView):
-    async def update_parents_recursive(self, id: Union[str, None], date: datetime) -> List[str]:
-        updated = []
-        
-        if id is not None:
-            updated.append(id)
-
-            await self.session.execute(update(Item).where(Item.id == id).values(date=date))
-            next_item = await self.session.execute(select(Item.parentId).where(Item.id == id))
-            next_item =next_item.scalars().first()
-
-            updated.extend(await self.update_parents_recursive(next_item, date))
-
-        return updated
-
-    async def insert_item(self, item: dict, date: datetime, update_parents: bool=False) -> List[str]:
+    async def insert_item_recursive(self, item: dict, date: datetime, update_parents: bool=False) -> List[str]:
         id = item['id']
-
-        await self.session.merge(create_item(item, date))
-
-        new_children = []
-        for child in item['children']:
-            new_children.extend(await self.insert_item(child, date))
-
-        self.session.add_all(map(lambda child: Relation(parentId=id, childId=child), new_children))
 
         row = await self.session.execute(select(Item.type, Item.parentId).where(Item.id == id))
         entry = row.first()
+
+        if entry is not None:
+            await self.session.execute(update(Item).where(Item.id == id).values(**dict(create_item(item, date))))
+        else:
+            await self.session.execute(insert(Item).values(**dict(create_item(item, date))))
+
+        new_children = []
+        for child in item['children']:
+            new_children.extend(await self.insert_item_recursive(child, date))
+
+        self.session.add_all(map(lambda child: Relation(parentId=id, childId=child), new_children))
 
         db_item_type, db_item_parent = None, None
 
@@ -64,17 +68,23 @@ class ImportsView(BaseView):
 
             if db_item_type != ItemType(item['type']):
                 raise ValidationError("unable to change item's type")
-            updated_parents = await self.update_parents_recursive(db_item_parent, date)
+            if (await self.session.execute(select(exists(Relation).where(and_(Relation.parentId == id, Relation.childId == item['parentId']))))).scalar_one():
+                raise ValidationError("invalid schema")
+            updated_parents = await update_parents_recursive(self.session, db_item_parent, date)
             if db_item_parent != item['parentId']:
                 new_children.append(id)
+                rows = await self.session.execute(select(Relation.childId).where(Relation.parentId==id))
+                new_children.extend(rows.scalars())
+
                 for parent in updated_parents:
-                    to_del = await self.session.get(Relation, (parent, id))
-                    await self.session.delete(to_del)
+                    for c in new_children:
+                        to_del = await self.session.get(Relation, (parent, c))
+                        await self.session.delete(to_del)
         else:
             new_children.append(id)
-
+            
         if update_parents and db_item_parent != item['parentId']:
-            updated_parents = await self.update_parents_recursive(item['parentId'], date)
+            updated_parents = await update_parents_recursive(self.session, item['parentId'], date)
             for parent in updated_parents:
                 self.session.add_all(map(lambda child: Relation(parentId=parent, childId=child), new_children))
 
@@ -109,7 +119,7 @@ class ImportsView(BaseView):
 
             async with session.begin():
                 for item in items:
-                    await self.insert_item(item, data['updateDate'], True)
+                    await self.insert_item_recursive(item, data['updateDate'], True)
                 await session.commit()
 
         await self.engine.dispose()
